Analyse the robust regression results by providing a table with descriptives

Analyse the robust regression results by providing a table with descriptives

analyse_meta(coeff, t_values)

Arguments

coeff
A dataframe of coefficients

t_values
A dataframe of t-values

Value

A dataframe with the descriptives

Examples

result <- rrs(mpg ~., data = mtcars, k = 4)
coeff <- result[1]
t_values <- result[2]
analyse_meta(coeff, t_values)

#>                 Mean Standard Deviation Conf. int. left Conf. int. right
#> am        3.42920475         1.61119888     -0.19628916      6.815592506
#> drat      1.99511519         1.13794133     -0.10530561      5.047654828
#> constant 25.79460751        12.71203434    -18.57373030     49.227796684
#> vs        1.35475076         1.51617586     -1.71713406      3.878364999
#> gear      0.96709003         1.52344995     -1.65997419      4.159209030
#> qsec      0.24529670         0.68214107     -1.50632230      1.228280039
#> disp     -0.01469883         0.01224722     -0.03771844      0.007869924
#> hp       -0.03109666         0.01412048     -0.06605174      0.000654710
#> carb     -1.07819617         0.51453298     -2.30065124     -0.146198317
#> cyl      -1.46923069         0.59591062     -3.07307257     -0.436979394
#> wt       -3.50908068         0.69385011     -5.14949505     -2.286573031
#>          Percent Significant Percent Neg. Significant Percent Pos. Significant
#> am                 0.5000000              0.000000000                0.5000000
#> drat               0.1547619              0.000000000                0.1547619
#> constant           0.7476190              0.004761905                0.7428571
#> vs                 0.1071429              0.000000000                0.1071429
#> gear               0.1666667              0.000000000                0.1666667
#> qsec               0.2023810              0.035714286                0.1666667
#> disp               0.3928571              0.392857143                0.0000000
#> hp                 0.6071429              0.607142857                0.0000000
#> carb               0.6190476              0.619047619                0.0000000
#> cyl                0.4761905              0.476190476                0.0000000
#> wt                 0.9761905              0.976190476                0.0000000
#>          Percent Positive Strong extreme bounds weak extreme bound test
#> am             0.98809524                     0                       0
#> drat           0.97619048                     0                       0
#> constant       0.96666667                     0                       0
#> vs             0.78571429                     0                       0
#> gear           0.66666667                     0                       0
#> qsec           0.61904762                     0                       0
#> disp           0.19047619                     0                       0
#> hp             0.03571429                     0                       0
#> carb           0.01190476                     0                       0
#> cyl            0.00000000                     0                       0
#> wt             0.00000000                     0                       1